pyproject.toml
[tool.poetry]
name = "zadanie-6105"
version = "0.1.0"
description = ""
authors = ["Komanok-dev <s.komanok@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.114.1"
sqlalchemy = "^2.0.34"
uvicorn = "^0.30.6"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
pydantic-settings = "^2.5.2"
python-jose = "^3.3.0"
passlib = "^1.7.4"
pytest = "^8.3.3"
httpx = "^0.27.2"
requests = "^2.32.3"
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
